crypto_trading_system/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-binance>=1.0.19
pandas>=2.1.0
websockets>=12.0
python-dotenv>=1.0.0
rich>=13.5.0
